refactor: split fullscreen overlay renderers into their own file
Move the fullscreen overlay renderers and the trailing case arms of
renderOverlayContentExtended out of view_overlays.go into a new
view_overlays_fullscreen.go, keeping view_overlays.go under the
800-line cap and reducing renderOverlayContentExtended's gocyclo.
